Move test_script.py into indexdigest/test/cli
